move the basic Jinja conf to `liberapay.website`
