+++ liberapay/renderers/jinja2.py
@@ -2,7 +2,7 @@ import aspen_jinja2_renderer as base
 
 from markupsafe import escape as htmlescape
 
-from liberapay.constants import JINJA_ENV_COMMON
+from liberapay.website import JINJA_ENV_COMMON
 
 
 class Renderer(base.Renderer):
